Move pages to their own folders and other shared components to shared
