+++ env.js
@@ -1,4 +1,4 @@
-// const EXPO_LOCAL_URL = '10.0.0.199' // Josh
-const EXPO_LOCAL_URL = '192.168.0.9' // Tia
+const EXPO_LOCAL_URL = '10.0.0.199' // Josh
+// const EXPO_LOCAL_URL = '192.168.0.9' // Tia
 
 export { EXPO_LOCAL_URL };
