--- src/components/CrawlingBugs/CrawlingBugs.js
@@ -1,16 +1,0 @@
-import { memo } from "react";
-import CrawlingBug from "./CrawlingBug";
-import { blue } from "../../styles/colors";
-
-const CrawlingBugs = memo(({ num = 8 }) => {
-  return Array(num)
-    .fill(1)
-    .map((_, i) => (
-      <CrawlingBug
-        key={i}
-        color={blue.hex}
-      />
-    ));
-});
-
-export default CrawlingBugs;
